+++ tools/scripts/cli_otd_process_gtfs.sh
@@ -1,7 +1,5 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-LOGS_BASEPATH=$DIR/logs
 
-# DATE_NOW=$(date +"%Y-%m-%d-%H%M")
 DATE_NOW=$(date +"%Y-%m-%d")
 
 source $DIR/common.sh