+++ tools/atlas-geocode-lines/CHANGELOG.md
@@ -4,5 +4,8 @@ This tool used in [Atlas-Lines compare GTFS-Routes](https://tools.odpch.ch/atlas
 
 ----
 
+10.March.2026
+- use and adapt code to use latest OJP SDK
+
 1.February.2025
 - added tool to `showcases` monorepo