+++ tools/gtfs-rt-static-compare/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG gtfs-rt-static-compare/
 
+10.Mar 2026
+- use and adapt code to use latest OJP SDK
+
 08.Feb 2025
 - adds compare info, check prev 10 measurements
 - send emails in case of errors detected